# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"admin",
							"staff",
							"user"
						]
					},
					"branchId": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string",
						"nullable": true
					},
					"isBusinessError": {
						"type": "boolean"
					},
					"errorDetail": {
						"type": "string",
						"nullable": true
					},
					"resultApi": {},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"isBusinessError",
					"errorDetail",
					"resultApi",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotifyType": {
				"type": "string",
				"enum": [
					"market",
					"food"
				]
			},
			"NotifyLevel": {
				"type": "string",
				"enum": [
					"global",
					"branch",
					"user"
				]
			},
			"NotifyDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/NotifyType"
					},
					"level": {
						"$ref": "#/components/schemas/NotifyLevel"
					},
					"createdBy": {
						"type": "string"
					},
					"branchId": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"readBy": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"message",
					"type",
					"level",
					"createdBy",
					"readBy",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateNotifyDto": {
				"properties": {
					"title": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/NotifyType"
					},
					"level": {
						"$ref": "#/components/schemas/NotifyLevel"
					},
					"branchId": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"message",
					"type",
					"level"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "sever-base-ctm",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					}
				}
			}
		},
		"/notifies": {
			"post": {
				"operationId": "CreateNotify",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/NotifyDto"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Notify"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateNotifyDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetNotifies",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/NotifyDto"
											},
											"type": "array"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Notify"
				],
				"security": [],
				"parameters": []
			}
		},
		"/notifies/{id}": {
			"get": {
				"operationId": "GetNotifyById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/NotifyDto"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Notify"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/notifies/{id}/read": {
			"post": {
				"operationId": "MarkAsRead",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/NotifyDto"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Notify"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}